load_stats: add reconcile to load_stats
This change adds the ability to reconcile the tablet sizes in
load_stats. This is needed because the tablet sizes in load_stats can be
stale: a merge or a split might have occured. Also, after a tablet migration
the tablet size must be moved in load_stats from source host to destination.